separate stream data structures from code
create lib to share code and structures between probe and monitor.
