+++ Cargo.toml
@@ -11,6 +11,9 @@ authors = ["Chris Kennedy"]
 version = "0.2.1"
 edition = "2021"
 
+[lib]
+path = "src/lib.rs"
+
 [features]
 default = []
 dpdk_enabled = ["capsule"]
